Add files via upload
The above code is programmed using python programming language. This code uses a huggingface API key and postman application can be used for making this code function without a front-end application. The smart email assistant fetches, categorizes and helps giving a reply according to the mail you have received with the help of AI.(The code is still under development for more optimized perfomance)